number combination app/news/bot/requirements.txt
feedparser==6.0.10
requests==2.31.0
beautifulsoup4==4.12.3
transformers==4.37.2
torch==2.2.0
Pillow==10.2.0
python-dotenv==1.0.1
supabase==2.10.0
nltk==3.8.1
lxml==5.1.0
tweepy==4.14.0
numpy<2
python-dateutil
